requirements.txt
opencv-python
numpy
scikit-image
scikit-learn
torch
torchvision
streamlit
Pillow
joblib